Move codycontext into frontend
Only used there.

Test plan:

Just moved code, CI will find issues.
