+++ cmd/frontend/internal/context/BUILD.bazel
@@ -7,9 +7,9 @@ go_library(
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/enterprise",
+        "//cmd/frontend/internal/codycontext",
         "//cmd/frontend/internal/context/resolvers",
         "//internal/codeintel",
-        "//internal/codycontext",
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/embeddings",
